refac: padatalistcomp style
