+++ test-web/src/components/DataListView/style/padatalistcompstyle.js
@@ -1,0 +1,27 @@
+const style = {
+    listContainer: {
+      textAlign: 'center',
+      width: '100%',
+      paddingRight: '0px',
+      paddingBottom: '0',
+      height: 'auto',
+      minWidth: '640px',
+    },
+    paginationBar: {
+      marginTop: '40px',
+      width: '100%',
+      justifyContent: 'center',
+    },
+    paginationContainer: {
+      display: 'flex',
+      justifyContent: 'right',
+      alignItems: 'center',
+      minWidth: '640px',
+    },
+    formControl: {
+      minWidth: 120,
+      marginLeft: '20px',
+    },
+  };
+
+  export default style
